+++ src/PDO/Connection/Php80.php
@@ -68,6 +68,10 @@ class Php80 extends Base
 			//  logs Error and throws SQL Exception
 			$this->logError( $e, $query );
 		}
+		finally{
+			if( static::LOG_LEVEL_UNSPECIFIED !== $this->logLevelForNextStatement )					//  one-time log level is set
+				$this->logLevelForNextStatement	= static::LOG_LEVEL_UNSPECIFIED;					//  reset, even if the query failed
+		}
 		return FALSE;
 	}
 }